backend/requirements.txt
fastapi
uvicorn
httpx
sqlalchemy
psycopg2-binary
python-docx
pdfplumber
python-multipart
fastapi.middleware.cors
python-dotenv